+++ src/libs/Battle_interface/sea/battle_shipsign.h
@@ -33,7 +33,7 @@ class BIShipIcon
     void Draw();
     void Init(ATTRIBUTES *pRoot, ATTRIBUTES *pA);
 
-    long AddTexture(const char *pcTextureName, long nCols, long nRows) const;
+    size_t AddTexture(const char *pcTextureName, long nCols, long nRows) const;
 
     void Recollect();
     void SetUpdate()
